backend/requirements.txt
fastapi==0.110.3
starlette==0.37.2
pydantic[email]==1.10.26
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
pymysql==1.1.2

python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9

numpy==1.26.4
opencv-python==4.10.0.84

pandas==2.2.3
openpyxl==3.1.5

# Optional face recognition (can be hard on Windows)
# face-recognition==1.3.0
